Fixes import issues after layout components have moved
