+++ app/_layout.tsx
@@ -4,7 +4,7 @@ import { StatusBar } from "expo-status-bar";
 import { BottomSheetModalProvider } from "@gorhom/bottom-sheet";
 import { GestureHandlerRootView } from "react-native-gesture-handler";
 
-import { TouchableOpacity, useColorScheme } from "react-native";
+import { TouchableOpacity } from "react-native";
 import { Ionicons } from "@expo/vector-icons";
 
 export const unstable_settings = {
@@ -14,7 +14,6 @@ export const unstable_settings = {
 
 export default function RootLayoutNav() {
   const navigation = useNavigation();
-  const colorScheme = useColorScheme();
 
   return (
     <GestureHandlerRootView style={{ flex: 1 }}>
